requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.6
django-heroku==0.3.1
dnspython==1.16.0
gunicorn==20.1.0
idna==3.2
ntlm-auth==1.5.0
psycopg2==2.9.1
pycparser==2.20
pymongo==3.12.0
python-snappy==0.6.0
pytz==2021.1
pywinrm==0.4.2
requests==2.26.0
requests-ntlm==1.1.0
#six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
#whitenoise==5.3.0
#winkerberos==0.8.0
#xmltodict==0.12.0
